+++ aggregator/utils.py
@@ -23,6 +23,7 @@
 from __future__ import annotations
 
 import logging
+import pickle
 import random
 import string
 import time
@@ -270,6 +271,15 @@ SIMPLEJSON_KWARGS = {
     'for_json': True,
     'iterable_as_array': True,
 }
+
+
+class RenamingUnpickler(pickle.Unpickler):
+    def find_class(self, module, name):
+        if module[:7] == 'feedly.':
+            module = 'aggregator.' + module[7:]
+        if name == 'ProbeRequest':
+            name = 'ProbeFeed'
+        return super().find_class(module, name)
 
 
 class HyperlinkStore(KeywordStore):