+++ presets/livejournal.py
@@ -1,7 +1,6 @@
 FOLLOW_DOMAINS = {'livejournal.com'}
 
-
-FEED_TEMPLATES = {
+RSS_TEMPLATES = {
     r'.*\.livejournal\.com/?.*': {
         'http://%(netloc)s/data/rss': 100,
         'https://%(netloc)s/data/rss': 200,